fix: durationMs Long 타입 변경

+++ src/main/java/ivory/ivory_be/runner/domain/RunnerEventMessageDto.java
@@ -24,7 +24,7 @@ public class RunnerEventMessageDto {
 
         private String line;
 
-        private Integer durationMs;
+        private Long durationMs;
 
         private Result result;
